[OPIK-2281] custom invite sidebar icon
