+++ apps/opik-frontend/src/components/layout/SideBar/SideBar.tsx
@@ -1,5 +1,4 @@
-import React, { MouseEventHandler, useState } from "react";
-import isNumber from "lodash/isNumber";
+import React, { useState } from "react";
 import { Link } from "@tanstack/react-router";
 
 import {
@@ -8,7 +7,6 @@ import {
   FlaskConical,
   GraduationCap,
   LayoutGrid,
-  LucideIcon,
   MessageCircleQuestion,
   FileTerminal,
   LucideHome,
@@ -30,7 +28,6 @@ import useOptimizationsList from "@/api/optimizations/useOptimizationsList";
 import { OnChangeFn } from "@/types/shared";
 import { Button } from "@/components/ui/button";
 import { Separator } from "@/components/ui/separator";
-import TooltipWrapper from "@/components/shared/TooltipWrapper/TooltipWrapper";
 import { buildDocsUrl, cn } from "@/lib/utils";
 import Logo from "@/components/layout/Logo/Logo";
 import usePluginsStore from "@/store/PluginsStore";
@@ -38,28 +35,13 @@ import ProvideFeedbackDialog from "@/components/layout/SideBar/FeedbackDialog/Pr
 import usePromptsList from "@/api/prompts/usePromptsList";
 import QuickstartDialog from "@/components/pages-shared/onboarding/QuickstartDialog/QuickstartDialog";
 import GitHubStarListItem from "@/components/layout/SideBar/GitHubStarListItem/GitHubStarListItem";
+import SidebarMenuItem, {
+  MENU_ITEM_TYPE,
+  MenuItem,
+  MenuItemGroup,
+} from "@/components/layout/SideBar/MenuItem/SidebarMenuItem";
 
-export enum MENU_ITEM_TYPE {
-  link = "link",
-  router = "router",
-  button = "button",
-}
-
-export type MenuItem = {
-  id: string;
-  path?: string;
-  type: MENU_ITEM_TYPE;
-  icon: LucideIcon;
-  label: string;
-  count?: string;
-  onClick?: MouseEventHandler<HTMLButtonElement>;
-};
-
-type MenuItemGroup = {
-  id: string;
-  label?: string;
-  items: MenuItem[];
-};
+type BottomMenuItem = MenuItem | (() => React.ReactNode);
 
 const HOME_PATH = "/$workspaceName/home";
 
@@ -175,57 +157,6 @@ type SideBarProps = {
   setExpanded: OnChangeFn<boolean | undefined>;
 };
 
-interface GetItemElement {
-  item: MenuItem;
-  content: React.ReactElement;
-  workspaceName: string;
-  linkClasses: string;
-}
-
-const getItemElementByType = ({
-  item,
-  content,
-  workspaceName,
-  linkClasses,
-}: GetItemElement) => {
-  if (item.type === MENU_ITEM_TYPE.router) {
-    return (
-      <li key={item.id} className="flex">
-        <Link to={item.path} params={{ workspaceName }} className={linkClasses}>
-          {content}
-        </Link>
-      </li>
-    );
-  }
-
-  if (item.type === MENU_ITEM_TYPE.link) {
-    return (
-      <li key={item.id} className="flex">
-        <a
-          href={item.path}
-          target="_blank"
-          rel="noreferrer"
-          className={linkClasses}
-        >
-          {content}
-        </a>
-      </li>
-    );
-  }
-
-  if (item.type === MENU_ITEM_TYPE.button) {
-    return (
-      <li key={item.id} className="flex">
-        <button onClick={item.onClick} className={cn(linkClasses, "text-left")}>
-          {content}
-        </button>
-      </li>
-    );
-  }
-
-  return null;
-};
-
 const SideBar: React.FunctionComponent<SideBarProps> = ({
   expanded,
   setExpanded,
@@ -238,9 +169,9 @@ const SideBar: React.FunctionComponent<SideBarProps> = ({
     setQuickstartOpened,
   } = useAppStore();
   const LogoComponent = usePluginsStore((state) => state.Logo);
-  // const sidebarMenuItems = usePluginsStore(
-  //   (state) => state.useSidebarMenuItems,
-  // )();
+  const SidebarInviteDevButton = usePluginsStore(
+    (state) => state.SidebarInviteDevButton,
+  );
 
   const { data: projectData } = useProjectsList(
     {
@@ -323,7 +254,14 @@ const SideBar: React.FunctionComponent<SideBarProps> = ({
     optimizations: optimizationsData?.total,
   };
 
-  const bottomMenuItems: MenuItem[] = [
+  const inviteDevButton = () => {
+    if (!SidebarInviteDevButton) {
+      return null;
+    }
+    return <SidebarInviteDevButton expanded={expanded} />;
+  };
+
+  const bottomMenuItems: BottomMenuItem[] = [
     {
       id: "documentation",
       path: buildDocsUrl(),
@@ -338,7 +276,7 @@ const SideBar: React.FunctionComponent<SideBarProps> = ({
       label: "Quickstart guide",
       onClick: () => setQuickstartOpened(true),
     },
-    // ...sidebarMenuItems,
+    inviteDevButton,
     {
       id: "provideFeedback",
       type: MENU_ITEM_TYPE.button,
@@ -354,45 +292,18 @@ const SideBar: React.FunctionComponent<SideBarProps> = ({
     <Logo expanded={expanded} />
   );
 
-  const renderItems = (items: MenuItem[]) => {
+  const renderItems = (items: BottomMenuItem[]) => {
     return items.map((item) => {
-      const hasCount = item.count && isNumber(countDataMap[item.count]);
-      const count = hasCount ? countDataMap[item.count!] : "";
-
-      const content = (
-        <>
-          <item.icon className="size-4 shrink-0" />
-          {expanded && (
-            <>
-              <div className="ml-1 grow truncate">{item.label}</div>
-              {hasCount && (
-                <div className="h-6 shrink-0 leading-6">{count}</div>
-              )}
-            </>
-          )}
-        </>
-      );
-
-      const linkClasses = cn(
-        "comet-body-s flex h-9 w-full items-center gap-2 text-foreground rounded-md hover:bg-primary-foreground data-[status=active]:bg-primary-100 data-[status=active]:text-primary",
-        expanded ? "pl-[10px] pr-3" : "w-9 justify-center",
-      );
-
-      const itemElement = getItemElementByType({
-        item,
-        content,
-        workspaceName,
-        linkClasses,
-      });
-
-      if (expanded) {
-        return itemElement;
+      if (typeof item === "function") {
+        return item();
       }
-
       return (
-        <TooltipWrapper key={item.id} content={item.label} side="right">
-          {itemElement}
-        </TooltipWrapper>
+        <SidebarMenuItem
+          key={item.id}
+          item={item}
+          expanded={expanded}
+          count={countDataMap[item.count!]}
+        />
       );
     });
   };
