+++ apps/opik-frontend/src/store/PluginsStore.ts
@@ -4,7 +4,7 @@ import { create } from "zustand";
 import WorkspacePreloader from "@/components/shared/WorkspacePreloader/WorkspacePreloader";
 import { GoogleColabCardCoreProps } from "@/components/pages-shared/onboarding/GoogleColabCard/GoogleColabCardCore";
 import { InviteDevButtonProps } from "@/plugins/comet/InviteDevButton";
-import { MenuItem } from "@/components/layout/SideBar/SideBar";
+import { SidebarInviteDevButtonProps } from "@/plugins/comet/SidebarInviteDevButton";
 
 type PluginStore = {
   Logo: React.ComponentType<{ expanded: boolean }> | null;
@@ -17,9 +17,9 @@ type PluginStore = {
     onChangeHeight: (height: number) => void;
   }> | null;
   InviteDevButton: React.ComponentType<InviteDevButtonProps> | null;
+  SidebarInviteDevButton: React.ComponentType<SidebarInviteDevButtonProps> | null;
   init: unknown;
   setupPlugins: (folderName: string) => Promise<void>;
-  useSidebarMenuItems: () => MenuItem[];
 };
 
 const VALID_PLUGIN_FOLDER_NAMES = ["comet"];
@@ -32,8 +32,8 @@ const PLUGIN_NAMES = [
   "WorkspacePreloader",
   "RetentionBanner",
   "InviteDevButton",
+  "SidebarInviteDevButton",
   "init",
-  "useSidebarMenuItems",
 ];
 
 const usePluginsStore = create<PluginStore>((set) => ({
@@ -45,7 +45,7 @@ const usePluginsStore = create<PluginStore>((set) => ({
   WorkspacePreloader: null,
   RetentionBanner: null,
   InviteDevButton: null,
-  useSidebarMenuItems: () => [],
+  SidebarInviteDevButton: null,
   init: null,
   setupPlugins: async (folderName: string) => {
     if (!VALID_PLUGIN_FOLDER_NAMES.includes(folderName)) {
